Move more traits stuff from ViewLegacy to ViewTraits
